When making a map iterator over a directory, calls to `.Next()` currently return a value of the `FieldHash()` of the dagpb link.
This change has them instead return a `IterLink`.

In both cases, the "exposed" interface of the returned type, `AsLink()` exposes the underlying hash cidLink to the entry.

By using a wrapping `IterLink` type, we provide an opportunity to type-cast the response, and instead of treating the response under it's returned `ipld.Node` interface,
the client has the ability to do something of the form:
```
k, next, err := iterator.Next()
nextSize := next.(*iter.IterLink).Substrate.FieldSize()
```